refacotr: recommendation Controller, Service 합치기
