+++ support-service/src/main/java/com/node5/supportservice/recommendation/presentation/RecommendationController.java
@@ -1,7 +1,10 @@
 package com.node5.supportservice.recommendation.presentation;
 
 import com.node5.supportservice.recommendation.application.RecommendationService;
+import com.node5.supportservice.recommendation.application.dto.ProductRecommendationInfo;
 import com.node5.supportservice.recommendation.application.dto.RecommendationResponse;
+import com.node5.supportservice.recommendation.presentation.dto.ProductRecommendationRequest;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,6 +19,18 @@ public class RecommendationController {
 
     private final RecommendationService recommendationService;
 
+    @Operation(summary = "상품 추천", description = "사용자 취향 임베딩을 기반으로 상품을 추천한다.")
+    @PostMapping
+    public ResponseEntity<ProductRecommendationInfo> recommend(@RequestBody ProductRecommendationRequest request) {
+        ProductRecommendationInfo recommendationInfo = recommendationService.recommendProducts(
+                request.preferenceEmbedding(),
+                request.excludedProductIds(),
+                request.limit()
+        );
+        return ResponseEntity.ok(recommendationInfo);
+    }
+
+    // TODO: 리네이밍
     @PostMapping
     public ResponseEntity<RecommendationResponse> recommend(
             @RequestHeader("Member-Id") UUID memberId,
